Add stable ids to events and enable filtering

+++ operator/hack/test/run-e2e-tests.sh
@@ -147,25 +147,21 @@ function run_proxy_checks() {
 }
 
 function run_k8s_events_checks() {
-  printf "Running k8s_events checks ..."
-
-  sleep 30
   # trigger events to be sent to the test-proxy
-  kubectl --namespace "$NS" rollout restart deploy wavefront-controller-manager &>/dev/null
-  sleep 10
+  wait_for_cluster_ready
+  "$REPO_ROOT/scripts/deploy/deploy-targets.sh"
+  wait_for_cluster_ready
 
   kubectl --namespace "$NS" port-forward "deploy/test-proxy" 8888 &>/dev/null &
   trap 'kill $(jobs -p) &>/dev/null || true' EXIT
-  sleep 3
 
   local external_events_results_file=$(mktemp)
   while true; do # wait until we get a good connection
-    RES_CODE=$(curl --silent --output "$external_events_results_file" --write-out "%{http_code}" "http://localhost:8888/events/external/assert")
+    RES_CODE=$(curl --silent --output "$external_events_results_file" --write-out "%{http_code}" "http://localhost:8888/events/external/assert" || echo "000")
     [[ $RES_CODE -lt 200 ]] || break
   done
 
   local external_event_count=$(jq ".EventCount" "$external_events_results_file")
-
 
   if [[ $external_event_count -eq 0 ]]; then
     red "missing external events."
@@ -173,7 +169,21 @@ function run_k8s_events_checks() {
     exit 1
   fi
 
-  echo " done."
+  local external_events_fail_count=$(jq "(.BadEventJSONs | length) + (.MissingFields | length) + (.FirstTimestampsMissing | length) + (.LastTimestampsInvalid | length)" "$external_events_results_file")
+
+  echo "$external_events_results_file"
+  if [[ external_events_fail_count -gt 0 ]]; then
+    red "BadEventJSONs: $(jq "(.BadEventJSONs | length)" "$external_events_results_file")"
+    red "MissingFields: $(jq "(.MissingFields | length)" "$external_events_results_file")"
+    red "FirstTimestampsMissing: $(jq "(.FirstTimestampsMissing | length)" "$external_events_results_file")"
+    red "LastTimestampsInvalid: $(jq "(.LastTimestampsInvalid | length)" "$external_events_results_file")"
+    if which pbcopy >/dev/null; then
+        echo "$external_events_results_file" | pbcopy
+      fi
+    exit 1
+  fi
+
+  "$REPO_ROOT/scripts/deploy/uninstall-targets.sh"
   kill $(jobs -p) &>/dev/null || true
 }
